git-svn-id: svn+ssh://jabouzi.com/home/jabouzic/svn/salatAndroid/trunk@119 86cb5e81-20db-4be9-97d3-054b39bd930a

+++ src/com/skanderjabouzi/salat/Location.java
@@ -7,7 +7,7 @@ public class Location {
 	private float longitude;
 	private String country;
 	private String city;
-	private int timezone;
+	private float timezone;
 
 	public long getId() {
 		return id;
@@ -33,11 +33,11 @@ public class Location {
 		this.longitude = longitude;
 	}
 
-	public int getTimezone() {
+	public float getTimezone() {
 		return timezone;
 	}
 
-	public void setTimezone(int timezone) {
+	public void setTimezone(float timezone) {
 		this.timezone = timezone;
 	}
